+++ CHANGELOG.md
@@ -25,10 +25,9 @@
 - Removed Boost Polling rate setting and changed the display itself to update when the entity is updated
 - Daily Streak HUD and Daily Challenge cards now format large progress numbers compactly
 - Some HUD elements now have buttons on inventory
-- Renamed Galatea to Moonglade Marsh
-- Made the emojis in party finder be sent in emoji form so it only occupies 1 char
 - Made party finder creation window save between sessions
 - Made panic keybind for custom fishing keybinds not disable the feature entirely.
+- Improved HUD editor snapping.
 
 ### Back-end
 - Made the mob tracking event driven, making it faster and improving Time to Kill accuracy.